requirements.txt
pillow >= 9.0
numpy
opencv-python
click